requirements.txt
aiosqlite==0.21.0
alembic==1.17.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
asyncpg==0.30.0
backports.asyncio.runner==1.2.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
coverage==7.11.0
cryptography==46.0.3
ecdsa==0.19.1
exceptiongroup==1.3.0
execnet==2.1.1
fastapi==0.120.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
paho-mqtt==2.1.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-env==1.2.0
pytest-xdist==3.8.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
redis==7.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
